Collapse entire public HTTP framework and experience fetch client into a single internal singleton ExperienceStore
