+++ Example/Example/AppDelegate.swift
@@ -16,7 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Pass your account token from the Rover Settings app to the Rover SDK.
-        Rover.accountToken = "<YOUR_SDK_TOKEN>"
+        Rover.accountToken = "d6ab40e8a45e3040c372806baba387fd"
         
         // Enable reporting of various Rover "events" such as when an experience is viewed or a button in a Rover
         // experience is tapped.
